requirements.txt
beautifulsoup4
fastapi
einops
func_argparse
html2image
jinja2
json_repair
jsonlines
lxml
mistune
marker-pdf==1.1.0
numpy
oaib
openai
opencv-python-headless
pandas
pdf2image
peft
pillow
PyPDF2
python-Levenshtein
python-multipart
python-pptx @ git+https://github.com/Force1ess/python-pptx
rich
socksio
tenacity
tiktoken
timm
transformers<4.50.0
uvicorn

